Fix broken options
In 6609e73 the changes I proposed were incorrectly extracted from our
application, and resulted in an error when the driver would be booted
for a spec.

This change, tested directly against our application, resolves the issue
by passing the right kind of options object (a custom options object
rather than an array of arguments) to `Selenium::WebDriver::Chrome`.

After merging, we should release version 0.4.4 and yank 0.4.3, as it is
not usable.